Meal App with Riverpod
Reduced the app complexity by reducing the function & data passing through unnecessary classes.  We use the riverpod library to manage the cross widget state in the app. So Now instead of passing the state through multiple classes  to eventually reach the desired location, we use riverpod library. Which is in other languages a dependency injection framework to reduce the tight coupling.